fix: change net month salary type

+++ src/main/java/com/devyforu/pibanks/Model/User.java
@@ -13,5 +13,5 @@ public class User {
     private String firstName;
     private String lastName;
     private Timestamp birthdayDate;
-    private Integer netMonthSalary;
+    private double netMonthSalary;
 }
